getting new access token and changing health data
