+++ app/src/main/java/edu/gatech/vera/vera/model/util/net/FitbitAPI.kt
@@ -4,6 +4,6 @@ object FitbitAPI {
     val clientID= "22BFC7"
     val clientSecret = "e81d8c0447eb0592f5e7e9d0b7e6905d"
     val callbackURI = "https%3A%2F%2Fnavaem8.wixsite.com"
-    val access_token = "eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIyMkJGQzciLCJzdWIiOiI3UE1NRzciLCJpc3MiOiJGaXRiaXQiLCJ0eXAiOiJhY2Nlc3NfdG9rZW4iLCJzY29wZXMiOiJyaHIgcmFjdCBycHJvIiwiZXhwIjoxNTc0MjU5ODA1LCJpYXQiOjE1NzQyMzEwMDV9.BvYHKEsACkK2QxInhd8YsGutHd_5J-QMflkwYF_uAN8"
+    var access_token = "eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIyMkJGQzciLCJzdWIiOiI3UE1NRzciLCJpc3MiOiJGaXRiaXQiLCJ0eXAiOiJhY2Nlc3NfdG9rZW4iLCJzY29wZXMiOiJyaHIgcmFjdCBycHJvIiwiZXhwIjoxNTc0MjU5ODA1LCJpYXQiOjE1NzQyMzEwMDV9.BvYHKEsACkK2QxInhd8YsGutHd_5J-QMflkwYF_uAN8"
     val refresh_token = "d9f29ff2b17a103c35f5643c6b0b2596dd7367764ad12545d6e143b14669495d"
 }
